# mvcc: remove defer in watchable store