Did profiling and found defer-call takes more resources in this case.