@gyuho 2016-05-19
BEFORE
```
go test -v -run=xxx -benchtime 3s -benchmem -cpuprofile prof.cpu -memprofile prof.mem -bench=BenchmarkWatchableStoreUnsyncedCancel
PASS
BenchmarkWatchableStoreUnsyncedCancel-8	 5000000	      1170 ns/op	       0 B/op	       0 allocs/op
ok  	github.com/coreos/etcd/mvcc	33.713s

ROUTINE ======================== github.com/coreos/etcd/mvcc.(*watchableStore).watch.func1 in /home/gyuho/go/src/github.com/coreos/etcd/mvcc/watchable_store.go
      50ms      3.42s (flat, cum)  7.81% of Total
         .          .    207:		slowWatcherGauge.Inc()
         .          .    208:		s.unsynced.add(wa)
         .          .    209:	}
         .          .    210:	watcherGauge.Inc()
         .          .    211:
      20ms       20ms    212:	cancel := cancelFunc(func() {
      30ms       70ms    213:		s.mu.Lock()
         .      150ms    214:		defer s.mu.Unlock()
         .          .    215:
         .          .    216:		// remove references of the watcher
         .      2.65s    217:		if s.unsynced.delete(wa) {
         .      110ms    218:			slowWatcherGauge.Dec()
         .       90ms    219:			watcherGauge.Dec()
         .      330ms    220:			return
         .          .    221:		}
         .          .    222:
         .          .    223:		if s.synced.delete(wa) {
         .          .    224:			watcherGauge.Dec()
         .          .    225:		}
```



AFTER

```
go test -v -run=xxx -benchtime 3s -benchmem -cpuprofile prof.cpu -memprofile prof.mem -bench=BenchmarkWatchableStoreUnsyncedCancel
PASS
BenchmarkWatchableStoreUnsyncedCancel-8	 5000000	      1078 ns/op	       0 B/op	       0 allocs/op
ok  	github.com/coreos/etcd/mvcc	31.139s


ROUTINE ======================== github.com/coreos/etcd/mvcc.(*watchableStore).watch.func1 in /home/gyuho/go/src/github.com/coreos/etcd/mvcc/watchable_store.go
      20ms      3.37s (flat, cum)  8.58% of Total
         .          .    207:		slowWatcherGauge.Inc()
         .          .    208:		s.unsynced.add(wa)
         .          .    209:	}
         .          .    210:	watcherGauge.Inc()
         .          .    211:
      10ms       10ms    212:	cancel := cancelFunc(func() {
      10ms       70ms    213:		s.mu.Lock()
         .          .    214:
         .          .    215:		// remove references of the watcher
         .      2.95s    216:		if s.unsynced.delete(wa) {
         .      140ms    217:			slowWatcherGauge.Dec()
         .      140ms    218:			watcherGauge.Dec()
         .       60ms    219:			s.mu.Unlock()
         .          .    220:			return
         .          .    221:		}
         .          .    222:
         .          .    223:		if s.synced.delete(wa) {
         .          .    224:			watcherGauge.Dec()

```

/cc @heyitsanthony @xiang90 

Please review.