@gyuho 2016-05-19
```
go test -v -run=xxx -benchtime 3s -benchmem -cpuprofile prof.cpu -memprofile prof.mem -bench=BenchmarkWatchableStoreUnsyncedCancel
PASS
BenchmarkWatchableStoreUnsyncedCancel-8	 5000000	      1005 ns/op	       0 B/op	       0 allocs/op
ok  	github.com/coreos/etcd/mvcc	32.474s


ROUTINE ======================== github.com/coreos/etcd/mvcc.(*watchableStore).watch.func1 in /home/gyuho/go/src/github.com/coreos/etcd/mvcc/watchable_store.go
      50ms      3.21s (flat, cum)  7.38% of Total
         .          .    208:		s.unsynced.add(wa)
         .          .    209:	}
         .          .    210:	watcherGauge.Inc()
         .          .    211:
         .          .    212:	cancel := cancelFunc(func() {
      20ms       50ms    213:		s.mu.Lock()
         .          .    214:		// remove references of the watcher
      20ms      2.90s    215:		if s.unsynced.delete(wa) {
         .       80ms    216:			slowWatcherGauge.Dec()
      10ms      140ms    217:			watcherGauge.Dec()
         .          .    218:		} else if s.synced.delete(wa) {
         .          .    219:			watcherGauge.Dec()
         .          .    220:		}
         .       40ms    221:		s.mu.Unlock()
         .          .    222:
         .          .    223:		// If we cannot find it, it should have finished watch.
         .          .    224:	})
         .          .    225:
         .          .    226:	return wa, cancel
```

@heyitsanthony PTAL.

Thanks.